my_learning_platform/requirements.txt
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
coverage==7.8.0
dj-database-url==2.3.0
Django==5.2.1
django-crispy-forms==2.4
django-extra-views==0.15.0
django-formtools==2.5.1
django-haystack==3.3.0
django-jazzmin==3.0.1
django-oscar==4.0
django-phonenumber-field==8.1.0
django-tables2==2.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.3.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
phonenumbers==9.0.5
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
purl==1.6
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
